==> insightful_routines_analysis/4_analyz_data.ipynb <==
import pandas as pd
import scipy.stats as stats


data = pd.read_csv("data_full.csv")

numerical_columns = data.select_dtypes(include=['int', 'float']).columns

correlation_results = {}
for column in numerical_columns:
    correlation_coef, p_value = stats.pearsonr(data[column], data['total_pimples'])
    correlation_results[column] = {'correlation_coef': correlation_coef, 'p_value': p_value}

sorted_results = sorted(correlation_results.items(), key=lambda x: x[1]['p_value'], reverse=True)
for column, result in sorted_results:
    print(f"Column: {column}, Correlation Coefficient: {result['correlation_coef']}, p-value: {result['p_value']}")

correlation = data.corr()["total_pimples"].sort_values(ascending=False)

print("Correlation with total_pimples:\n", correlation)
