==> insightful_routines_analysis/5_prediction.ipynb <==
import pandas as pd
import optuna

from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error

data = pd.read_csv("data_full.csv")

X = data.drop(columns=['total_pimples'])
y = data['total_pimples']

# def objective(trial):
#     n_estimators = trial.suggest_int('n_estimators', 100, 1000)
#     max_depth = trial.suggest_int('max_depth', 10, 50)
#     min_samples_split = trial.suggest_int('min_samples_split', 2, 32)
#     min_samples_leaf = trial.suggest_int('min_samples_leaf', 1, 32)
#     max_features = trial.suggest_categorical('max_features', ['sqrt', 'log2'])
#     criterion = trial.suggest_categorical('criterion', ["squared_error", "absolute_error", "friedman_mse", "poisson"])

#     model = RandomForestRegressor(
#         n_estimators=n_estimators,
#         max_depth=max_depth,
#         min_samples_split=min_samples_split,
#         min_samples_leaf=min_samples_leaf,
#         max_features=max_features,
#         criterion=criterion,
#         random_state= 21
#     )

#     model.fit(X_train, y_train)
#     y_pred = model.predict(X_test)

#     # metric  to optimize
#     score = mean_squared_error(y_test, y_pred)
    
#     return score

# study = optuna.create_study(direction='minimize', sampler=optuna.samplers.RandomSampler(seed=42))
# study.optimize(objective, n_trials=500)

# # Print the best parameters found 
# print("Best trial:")
# trial = study.best_trial

# print("Value: {:.4f}".format(trial.value))

# print("Params: ")
# for key, value in trial.params.items():
#     print("    {}: {}".format(key, value))

"""
    Best trial:
    Value: 2.6245
    Params: 
        n_estimators: 366
        max_depth: 19
        min_samples_split: 3
        min_samples_leaf: 1
        max_features: sqrt
        criterion: poisson
"""

params = {'n_estimators': 366, 'max_depth': 19, 'min_samples_split': 3, 'min_samples_leaf': 1, 'max_features': 'sqrt', 'criterion': 'poisson'}

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

model = RandomForestRegressor(**params)

model.fit(X_train, y_train)

predictions = model.predict(X_test)

# Evaluate the model
mse = mean_squared_error(y_test, predictions)
print("Mean Squared Error:", mse)

# Analyze feature importances
feature_importances = model.feature_importances_
print("Feature Importances:", feature_importances)
